Fix bug where fee from difficulty window was 0
The fee from a difficulty window could incorrectly return as zero
if the first window was the newest, incomplete window and there were
not any new tickets in that window yet. This has been corrected by
checking for the size of the first window, and also by skipping any
zero fee windows.